refactor: move network recovery logic into it's own module
